+++ Drivers/UserDrivers/DWINTFT/DWINTFT_Radar.h
@@ -10,14 +10,14 @@
 #define TFT_MUTE 0x05
 
 void TFT_SetProbeVersion(UART_HandleTypeDef *huart, uint8_t ProbeVersion);
-void TFT_SetRadarOrder(UART_HandleTypeDef *huart, uint8_t *pRadarOrder, uint8_t MaxProbeNum);
+void TFT_SetRadarOrder(UART_HandleTypeDef *huart, uint32_t *pRadarOrder, uint8_t MaxProbeNum);
 void TFT_ExchangeRadarOrder(UART_HandleTypeDef *huart, uint8_t n1, uint8_t n2, uint8_t MaxProbeNum);
 void TFT_DispRadarDist(UART_HandleTypeDef *huart, uint8_t RadarMinDist);
 void TFT_DispRadarColor(UART_HandleTypeDef *huart, uint8_t *pRadarColor, uint8_t MaxProbeNum);
 void TFT_DispVehicleSpeed(UART_HandleTypeDef *huart, uint8_t speed);
 void TFT_ReadProbeOrder(UART_HandleTypeDef *huart);
 
-extern uint8_t RadarProbeOrder[];
+extern uint32_t RadarProbeOrder[];
 extern uint32_t RadarLimitDist;
 
 #endif
